Created Population.cs and moved population logic in from GeneticSimulation Created PairedChromosomes.cs which contains a female + male chromosome pair Modified Genetic Simulation interfaces to implement actions based on new population and paired chromosome data structures and modified example classes to adhere to these new interfaces Begun creating universal UI in GeneticAlgorithmsUI project

+++ GeneticAlgorithmsUI/GenerationDataGridItem.cs
@@ -9,8 +9,8 @@ namespace GeneticAlgorithmsUI
     class GenerationDataGridItem
     {
         public string GenerationName { get; set; }
-        public float AverageFitness { get; set; }
-        public float MaximumFitness { get; set; }
+        public double AverageFitness { get; set; }
+        public double MaximumFitness { get; set; }
         public string MostSuccessfulChromosome { get; set; }
         public string LeastSuccessfulChromosome { get; set; }
     }
